苍穹外卖中间项目-04

+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -8,7 +8,6 @@ import com.sky.enumeration.OperationType;
 import com.sky.vo.DishVO;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
 
 import java.util.List;
 
@@ -17,7 +16,7 @@ public interface DishMapper {
 
     Page<DishVO> pageQuery(DishPageQueryDTO dishPageQueryDTO);
 
-    DishVO queryById(Integer id);
+    DishVO queryById(Long id);
 
     List<Dish> queryByDish(Dish dish);
